VCD: fixed error occurred while fetching accounts information - changed the field "McId" of class "AccountInfo" from datatype "int" to "long".

+++ 3/API/SeleniumDataBrowser/VCD/Helpers/UserInfoExtracting/Models/PageAccountInfo.cs
@@ -12,6 +12,6 @@
         public int primaryBusinessGroupId { get; set; }
         public string retailSubscriptionLevel { get; set; }
         public string status { get; set; }
-        public int mcId { get; set; }
+        public long mcId { get; set; }
     }
 }
